Changed continue button
Continue button now only appears when the timer of a set blind reaches 0.

Also added a play and pause button which fade in while time is being counted, and out in between rounds

Also adjusted the forms in the settings screen as were missing a bootstrap class.

Next actions are to change the javascript to set the blinds into an array