Polishes
Have amended pause and play buttons into one with a variable icon depending on whether it is currently counting down.
Moved the settings button to the top left to prevent overlap with the scroll bar
Have made the background image stay fixed in place and made the contents scroll instead - now extra blind rows won't show on top of a white background.